Fully Function GameContainer and Scene Switching
I created some placeholders for scenes that other people are currently working on.
I left notes (in the form of labels) of how to properly connect your scenes to the chain of things.
Please move and rename the scenes you make to replace the placeholder scenes.
There are a bunch of some resource methods in GameContainer. GameContainer is not a singleton, HOWEVER because of the setup it is always instantiated and it controls the scene order by instantiating new scenes as its children. ALSO you can access the game container at any time non-staticly with GameContainer.GAME_CONTAINER.

TODO: Option was left open for overlay panels (ie a pause menu). Did not implement yet.